+++ workalendar/exceptions.py
@@ -1,0 +1,9 @@
+"""
+Core Workalendar Exceptions
+"""
+
+
+class CalendarError(Exception):
+    """
+    Base Calendar Error
+    """